feat: remove unused RxSwift views
