--- p2p_wallet/UIKit/Views/WalletsCollectionView/WalletsCollectionView.swift
@@ -1,32 +1,0 @@
-//
-//  WalletsCollectionView.swift
-//  p2p_wallet
-//
-//  Created by Chung Tran on 26/03/2021.
-//
-
-import BECollectionView
-import Foundation
-import RxSwift
-
-class WalletsCollectionView: BEStaticSectionsCollectionView {
-    let walletsRepository: WalletsRepository
-
-    init(
-        header: BECollectionViewHeaderFooterViewLayout? = nil,
-        walletsRepository: WalletsRepository,
-        sections: [BEStaticSectionsCollectionView.Section],
-        footer: BECollectionViewHeaderFooterViewLayout? = nil
-    ) {
-        self.walletsRepository = walletsRepository
-        super.init(
-            header: header,
-            sections: sections,
-            footer: footer
-        )
-    }
-
-    override func dataDidChangeObservable() -> Observable<Void> {
-        walletsRepository.dataDidChange
-    }
-}
